requirements.txt
fastapi==0.112.0
uvicorn==0.30.0
line-bot-sdk==3.11.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
Pillow>=10.0